+++ packages/@leight/file/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @leight/file
 
+## 0.5.61
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/prisma@0.5.70
+
 ## 0.5.60
 
 ### Patch Changes