+++ packages/@leight/user-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @leight/user-server
 
+## 0.5.66
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/prisma@0.5.70
+    - @leight/user@0.5.64
+
 ## 0.5.65
 
 ### Patch Changes